test(parity): split workflow parity proof by service
Why:
- the workflow parity proof had grown into a single oversized test module that
  made failures hard to diagnose and parity harder to audit
- the repo standards call for smaller test seams so workflow behaviors stay
  directly inspectable instead of being buried in one aggregate file

What:
- replace the monolithic workflow parity module with service-focused workflow
  tests for manifesting, batch screening, batch staging, verification compare,
  and wallet-inventory rebuild behavior
- isolate duplicate-stage handling from normalization-window staging proof so
  each failure surface stays narrow
- update the parity ledger totals and engineering standards hotspot list to
  reflect the current suite and the structured CSV package split

Checks:
- uv run python -m tools.run_quality_gates --full-tests
